requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
lightfm==1.17
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
supabase==2.4.0
python-dotenv==1.0.1
python-multipart==0.0.9